Refactor AddAction structure and add saving of TOTP
