+++ Cargo.toml
@@ -33,3 +33,4 @@ hex = "0.4.3"
 keepass-ng = { path = "/Users/anssi/projects/keepass-rs", features = ["save_kdbx4", "totp"] }
 keyring = "2.3.2"
 uuid = { version = "1.8.0", features = ["v4"] }
+percent-encoding = "2.3.1"
